+++ doc/release-notes.md
@@ -2,7 +2,7 @@
 - upgrade to kind kubernetes 1.34
 - upgrade to postgres 18.1
 - removed kong and ingress definitions
-- upgrade to apsix chart 2.12.6 with ingress controller 2.0
+- upgrade to apsix chart 2.14.0 with ingress controller 2.0
 - switch to gateway api
 
 # 1.4.3